backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-binance==1.0.19
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3  # Pin to 4.1.3 for passlib compatibility
bcrypt==4.1.3
python-multipart==0.0.6
sqlalchemy>=2.0.35
# psycopg2-binary==2.9.9  # Optional - only needed for PostgreSQL
alembic==1.13.0
python-dotenv==1.0.0
aiosqlite==0.19.0
